requirements.txt
asgiref==3.8.1
boto3==1.38.22
botocore==1.38.22
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.0.9
django-cors-headers==4.4.0
django-environ==0.12.0
djangorestframework==3.15.2
idna==3.10
jmespath==1.0.1
mssql-django==1.5
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
waitress==3.0.2
whitenoise==6.8.2
openpyxl==3.1.5
minio==7.2.20